refactor(MatchResult): 패캐지 변경에 따른 수정 사항 전역 적용
