+++ src/main/java/view/OutputView.java
@@ -4,7 +4,7 @@ import java.util.List;
 import model.dto.DealerScoreResult;
 import model.dto.FaceUpResult;
 import model.dto.PlayerScoreResult;
-import model.dto.MatchResult;
+import model.casino.MatchResult;
 
 public class OutputView {
     private static final int HOLE_CARD_INDEX = 0;
